Add support for Emfatic imports
When breaking off the IFC part of the Hawk API, we needed to support the
case when part of an API might be imported from another .ecore file. To
do this, the generator has been changed in several ways:

- The services will be only the ones under the main EPackage (the one
  from the main .ecore file).

- The entities, exceptions and enums will be those under the main
  EPackage, and any from the imported EPackages that are referenced
  from the services in the main .ecore file.

This means that entities, exceptions and enums will be duplicated for now.
We may add support for the Thrift 'include' statement later, but this
would require specifying a VCS-friendly way of generating the path to the
'.thrift' files produced by the imported '.ecore' files.